refactor: BasicAuthenticationFilter를 추상화하라
