+++ src/main/java/nextstep/auth/authentication/AuthenticationToken.java
@@ -4,6 +4,9 @@ public class AuthenticationToken {
     private String principal;
     private String credentials;
 
+    public AuthenticationToken() {
+    }
+
     public AuthenticationToken(String principal, String credentials) {
         this.principal = principal;
         this.credentials = credentials;
