+++ src/main/java/nextstep/auth/converter/BasicAuthenticationConverter.java
@@ -8,12 +8,15 @@ import org.apache.tomcat.util.codec.binary.Base64;
 import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
 
-public class BasicAuthenticationConverter implements AuthenticationConverter{
+public class BasicAuthenticationConverter implements AuthenticationConverter {
     @Override
     public AuthenticationToken convert(HttpServletRequest request) throws IOException {
         String authCredentials = AuthorizationExtractor.extract(request, AuthorizationType.BASIC);
-        String authHeader = new String(Base64.decodeBase64(authCredentials));
+        if (authCredentials.isEmpty()) {
+            return new AuthenticationToken();
+        }
 
+        String authHeader = new String(Base64.decodeBase64(authCredentials));
         String[] splits = authHeader.split(":");
         String principal = splits[0];
         String credentials = splits[1];
